+++ docs/conf.py
@@ -101,4 +101,5 @@ intersphinx_mapping = {
     "https://docs.python.org/3/": None,
     "https://aiohttp.readthedocs.io/en/stable/": None,
     "https://aiohttp-middlewares.readthedocs.io/en/stable/": None,
+    "https://pyrsistent.readthedocs.io/en/stable/": None,
 }
